requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.2.1
defusedxml==0.7.1
duckduckgo_search==8.1.1
fonttools==4.59.0
fpdf2==2.8.3
frozenlist==1.5.0
idna==3.4
jplephem==2.22
lxml==6.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
numpy==2.2.3
openai==0.27.2
pillow==11.3.0
primp==0.15.0
propcache==0.2.1
pvrecorder==1.1.1
pygame==2.3.0
Pygments==2.19.2
PyMuPDF==1.25.5
pyperclip==1.8.2
python-dotenv==1.1.1
requests==2.28.2
rich==14.0.0
sgp4==2.24
skyfield==1.51
tqdm==4.67.1
urllib3==1.26.14
yarl==1.18.3
